requirements.txt
cassandra-driver
celery==3.1
redis
stream-framework[cassandra]
Flask
flask-restful
